pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Document Extraction API Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.0.0",
    "werkzeug>=3.0.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
]
